+++ simplestruct/__init__.py
@@ -3,7 +3,7 @@ similar to collections.namedtuple classes but support optional type-
 checking, mutability, and inheritance.
 """
 
-__version__ = '0.2.2-dev'
+__version__ = '0.2.2'
 
 from .struct import *
 from .fields import *